+++ src/ast/index.ts
@@ -4,6 +4,7 @@
 export * from './types.js';
 export * from './builder/index.js';
 export * from './printer/index.js';
+export * from './converter/index.js';
 export * from './stringify.js';
 export * from './create-element.js';
 export * from './parse.js';
