--- src/attribute/index.ts
@@ -1,7 +1,0 @@
-/**
- * @module ts-graphviz/attribute
- */
-export * from './attribute.js';
-export * from './keys.js';
-export * from './key-dict.js';
-export * from './C.js';
