+++ src/common/models.ts
@@ -1,4 +1,4 @@
-import type { Compass } from '../type/index.js';
+import type { Compass } from './type/index.js';
 import type {
   Attribute,
   AttributeKey,
@@ -7,15 +7,20 @@ import type {
   NodeAttributeKey,
   GraphAttributeKey,
   SubgraphAttributeKey,
-} from '../attribute/index.js';
-import type {
-  ASTNode,
-  NodeASTNode,
-  GraphASTNode,
-  AttributeListASTNode,
-  SubgraphASTNode,
-  EdgeASTNode,
-} from '../ast/index.js';
+} from './attribute/index.js';
+
+export type ASTType =
+  | 'Literal'
+  | 'Dot'
+  | 'Graph'
+  | 'Attribute'
+  | 'Comment'
+  | 'AttributeList'
+  | 'NodeRef'
+  | 'NodeRefGroup'
+  | 'Edge'
+  | 'Node'
+  | 'Subgraph';
 
 /**
  * Objects that can be Edge destinations satisfy this interface.
@@ -50,8 +55,8 @@ export type NodeAttributesObject = AttributesObject<NodeAttributeKey>;
 export type GraphAttributesObject = AttributesObject<GraphAttributeKey>;
 export type SubgraphAttributesObject = AttributesObject<ClusterSubgraphAttributeKey | SubgraphAttributeKey>;
 
-export interface Model<T extends ASTNode> {
-  $$type: T['type'];
+export interface ASTConvartable<T extends ASTType> {
+  $$type: T;
 }
 
 export interface HasComment {
@@ -86,7 +91,7 @@ export interface AttributeListModel<
   T extends AttributeKey = AttributeKey,
 > extends Attributes<T>,
     HasComment,
-    Model<AttributeListASTNode> {
+    ASTConvartable<'AttributeList'> {
   $$kind: K;
 }
 
@@ -95,12 +100,12 @@ export interface Port {
   compass: Compass;
 }
 
-export interface NodeModel extends HasComment, HasAttributes<NodeAttributeKey>, Model<NodeASTNode> {
+export interface NodeModel extends HasComment, HasAttributes<NodeAttributeKey>, ASTConvartable<'Node'> {
   readonly id: string;
   port(port: string | Partial<Port>): ForwardRefNode;
 }
 
-export interface EdgeModel extends HasComment, HasAttributes<EdgeAttributeKey>, Model<EdgeASTNode> {
+export interface EdgeModel extends HasComment, HasAttributes<EdgeAttributeKey>, ASTConvartable<'Edge'> {
   readonly targets: EdgeTargetTuple;
 }
 
@@ -453,11 +458,11 @@ export interface GraphBaseModel<T extends AttributeKey = AttributeKey> extends H
 
 export interface SubgraphModel
   extends GraphBaseModel<SubgraphAttributeKey | ClusterSubgraphAttributeKey>,
-    Model<SubgraphASTNode> {
+    ASTConvartable<'Subgraph'> {
   isSubgraphCluster(): boolean;
 }
 
-export interface RootGraphModel extends GraphBaseModel<GraphAttributeKey>, Model<GraphASTNode> {
+export interface RootGraphModel extends GraphBaseModel<GraphAttributeKey>, ASTConvartable<'Graph'> {
   directed: boolean;
 
   /**
@@ -470,4 +475,51 @@ export interface RootGraphModel extends GraphBaseModel<GraphAttributeKey>, Model
    * Subsequent edge statements using the same two nodes will identify the edge with the previously defined one and apply any attributes given in the edge statement.
    */
   strict: boolean;
+}
+
+export function isForwardRefNode(object: unknown): object is ForwardRefNode {
+  return typeof object === 'object' && object !== null && typeof (object as ForwardRefNode).id === 'string';
+}
+
+export function isNodeModel(object: unknown): object is NodeModel {
+  return (
+    typeof object === 'object' &&
+    object !== null &&
+    (object as NodeModel).$$type === 'Node' &&
+    typeof (object as NodeModel).id === 'string'
+  );
+}
+
+export function isNodeRef(node: unknown): node is NodeRef {
+  return isNodeModel(node) || isForwardRefNode(node);
+}
+
+export function isNodeRefLike(node: unknown): node is NodeRefLike {
+  return typeof node === 'string' || isNodeRef(node);
+}
+
+export function isNodeRefGroupLike(target: NodeRefLike | NodeRefGroupLike): target is NodeRefGroupLike {
+  return Array.isArray(target) && target.every(isNodeRefLike);
+}
+
+export function isCompass(c: string): c is Compass {
+  return ['n', 'ne', 'e', 'se', 's', 'sw', 'w', 'nw', 'c'].includes(c);
+}
+
+export function toNodeRef(target: NodeRefLike): NodeRef {
+  if (isNodeRef(target)) {
+    return target;
+  }
+  const [id, port, compass] = target.split(':');
+  if (isCompass(compass)) {
+    return { id, port, compass };
+  }
+  return { id, port };
+}
+
+export function toNodeRefGroup(targets: NodeRefGroupLike): NodeRefGroup {
+  if (targets.length < 2 && (isNodeRefLike(targets[0]) && isNodeRefLike(targets[1])) === false) {
+    throw Error('EdgeTargets must have at least 2 elements.');
+  }
+  return targets.map((t) => toNodeRef(t));
 }
