+++ src/index.ts
@@ -1,8 +1,7 @@
 /**
  * @module ts-graphviz
  */
-export { attribute } from './attribute/index.js';
-export * from './model/index.js';
-export * as AST from './ast/index.js';
+export * from './attribute.js';
+export * from './models.js';
 export * from './create-root-graph.js';
 export * from './to-dot.js';
