--- src/model/index.ts
@@ -1,6 +1,0 @@
-/**
- * @module ts-graphviz/model
- */
-export * from './ast-converter.js';
-export * from './models.js';
-export * from './types.js';
