+++ typedoc.json
@@ -2,12 +2,6 @@
   "out": "./doc",
   "entryPoints": [
     "src/index.ts",
-    "src/ast/index.ts",
-    "src/attribute/index.ts",
-    "src/model/index.ts",
-    "src/parser/index.ts",
-    "src/renderer/index.ts",
-    "src/type/index.ts",
-    "src/usecase/index.ts"
+    "src/ast/index.ts"
   ]
 }
